Fix cleaning stray pipelines from tests
