+++ internal/integration/integration_test.go
@@ -2,77 +2,16 @@ package integration_test
 
 import (
 	"context"
-	"embed"
 	"fmt"
-	"log"
-	"os"
-	"reflect"
-	"strings"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/buildkite/agent-stack-k8s/v2/api"
-	"github.com/buildkite/agent-stack-k8s/v2/cmd/controller"
-	"github.com/buildkite/agent-stack-k8s/v2/internal/controller/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-const (
-	repoHTTP = "https://github.com/buildkite/agent-stack-k8s"
-	repoSSH  = "git@github.com:buildkite/agent-stack-k8s"
-)
-
-var (
-	branch                  string
-	preservePipelines       bool
-	deleteOrphanedPipelines bool
-	cfg                     config.Config
-
-	//go:embed fixtures/*
-	fixtures embed.FS
-)
-
-// hacks to make --config work
-func TestMain(m *testing.M) {
-	if branch == "" {
-		log.Fatalf(
-			`You need to run the tests with a flag: -ldflags="-X %s.branch=$BRANCH_NAME"`,
-			reflect.TypeOf(testcase{}).PkgPath(),
-		)
-	}
-
-	if err := os.Chdir(".."); err != nil {
-		log.Fatal(err)
-	}
-	cmd := controller.New()
-	cmd.Flags().
-		BoolVar(&preservePipelines, "preserve-pipelines", false, "preserve pipelines created by tests")
-	cmd.Flags().
-		BoolVar(&deleteOrphanedPipelines, "delete-orphaned-pipelines", false, "delete all pipelines matching agent-k8s-*")
-
-	var err error
-	cfg, err = controller.ParseConfig(cmd, os.Args[1:])
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	if err := os.Chdir("integration"); err != nil {
-		log.Fatal(err)
-	}
-
-	for i, v := range os.Args {
-		if strings.Contains(v, "test") {
-			os.Args[i] = v
-		} else {
-			os.Args[i] = ""
-		}
-	}
-
-	os.Exit(m.Run())
-}
 
 func TestWalkingSkeleton(t *testing.T) {
 	tc := testcase{
@@ -276,13 +215,18 @@ func TestCleanupOrphanedPipelines(t *testing.T) {
 	if !deleteOrphanedPipelines {
 		t.Skip("not cleaning orphaned pipelines")
 	}
+
 	ctx := context.Background()
 	graphqlClient := api.NewClient(cfg.BuildkiteToken)
 
-	pipelines, err := api.SearchPipelines(ctx, graphqlClient, cfg.Org, "agent-k8s-", 100)
+	pipelines, err := api.SearchPipelines(ctx, graphqlClient, cfg.Org, "agent-stack-k8s-", 100)
 	require.NoError(t, err)
+
+	numPipelines := len(pipelines.Organization.Pipelines.Edges)
+	t.Logf("found %d pipelines to delete", numPipelines)
+
 	var wg sync.WaitGroup
-	wg.Add(len(pipelines.Organization.Pipelines.Edges))
+	wg.Add(numPipelines)
 	for _, pipeline := range pipelines.Organization.Pipelines.Edges {
 		pipeline := pipeline // prevent loop variable capture
 		t.Run(pipeline.Node.Name, func(t *testing.T) {
@@ -307,7 +251,7 @@ func TestCleanupOrphanedPipelines(t *testing.T) {
 				GraphQL: api.NewClient(cfg.BuildkiteToken),
 			}.Init()
 			tc.PipelineName = pipeline.Node.Name
-			tc.deletePipeline(context.Background())
+			tc.deletePipeline(ctx)
 		})
 	}
 }
@@ -324,10 +268,7 @@ func TestEnvVariables(t *testing.T) {
 	tc.StartController(ctx, cfg)
 	build := tc.TriggerBuild(ctx, pipelineID)
 	tc.AssertSuccess(ctx, build)
-	tc.AssertLogsContain(
-		build,
-		"Testing some env variables: set",
-	)
+	tc.AssertLogsContain(build, "Testing some env variables: set")
 }
 
 func TestImagePullBackOffCancelled(t *testing.T) {
